agent/pyproject.toml
[tool.poetry]
name = "dsa_agent"
version = "0.1.0"
description = "Starter"
authors = ["Arya Pratap Singh <notaryasingh@gmail.com>"]
license = "MIT"

[project]
name = "dsa_agent"
version = "0.0.1"

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.poetry.dependencies]
python = "^3.12"
langchain-anthropic = "^0.2.1"
langchain = "^0.3.1"
openai = "^1.51.0"
langchain-community = "^0.3.1"
copilotkit = "0.1.30"
uvicorn = "^0.31.0"
python-dotenv = "^1.0.1"
langchain-core = "^0.3.25"
langgraph-cli = {extras = ["inmem"], version = "^0.1.64"}
langchain-groq = "^0.2.2"
chroma = "^0.2.0"
fastembed = "^0.5.0"
bs4 = "^0.0.2"
faiss-cpu = "^1.9.0.post1"
ipykernel = "^6.29.5"

[tool.poetry.scripts]
demo = "dsa_agent.demo:main"